Commonize test utils
